[Breaking change] Improve import_id to avoid missing duplicate entries
Without this change, some memo's were different for the purchase when
downloading the statement in different moments. This makes the import_id's usage
of memo more conservative, picking only the first 4 characters.

Breaking: this will change the import_id generated, and so re-importing old
bills will end up ignoring what was already imported. If this is a problem for
you, add an if condition in the Ynab class such that only transaction after a
certain date are loaded.